% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_performance_fun.R
\name{benchmark_performance}
\alias{benchmark_performance}
\title{If you wanna test the performance of your model on synthetic generated test spots you can use this function to benchmark and get a sense of the model's performance.}
\usage{
benchmark_performance(ground_truth_mtrx, deconvoluted_mtrx)
}
\arguments{
\item{ground_truth_mtrx}{Object of class matrix containing the ground truth composition of each spot.}

\item{deconvoluted_mtrx}{Object of class matrix with the predicted cell type composition of each spot.}
}
\value{
This function returns a list with RMSE, RMSE_ct, JSD, PCC and PCC_ct.
}
\description{
If you wanna test the performance of your model on synthetic generated test spots you can use this function to benchmark and get a sense of the model's performance.
}
