% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast.sc.cell.label}
\alias{breast.sc.cell.label}
\title{cell type labels of scRNA-seq data}
\format{
a vector
}
\usage{
data(breast.sc.cell.label)
}
\description{
cell type labels of scRNA-seq data
}
\examples{
data(breast.sc.cell.label)

}
\keyword{datasets}
