% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast.sc.ref}
\alias{breast.sc.ref}
\title{raw count matrix of reference scRNA-seq dataset}
\format{
a large matrix
}
\usage{
data(breast.sc.ref)
}
\description{
We obtain the filtered human breast cancers scRNA-seq dataset from
[Zenodo data repository] (<https://doi.org/10.5281/zenodo.4739739>)
The dataset contains the expression levels of 11920 genes and 3024 cells.
}
\examples{
data(breast.sc.ref)

}
\keyword{datasets}
