% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast.spot.annotation}
\alias{breast.spot.annotation}
\title{The annotation of spots of breast cancer data}
\format{
a vector
}
\usage{
data(breast.spot.annotation)
}
\description{
109 spots in connective tissue, 7 spots in immune infiltrate, 139 spots in invasive cancer, 51 spots in undermined.
The breast.spot.annotation just for the down-streaming analysis.
}
\examples{
data(breast.spot.annotation)

}
\keyword{datasets}
