% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast.st.loc}
\alias{breast.st.loc}
\title{coordinate of spots for the spatially resolved transcriptomics data}
\format{
a list
}
\usage{
data(breast.st.loc)
}
\description{
coordinate of spots for the breast.st. The center of the grids are served as the coordinates of the corresponding generated spots.
}
\examples{
data(breast.st.loc)

}
\keyword{datasets}
