% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast.st}
\alias{breast.st}
\title{raw count matrix of spatial transcriptomics data}
\format{
a large matrix
}
\usage{
data(breast.st)
}
\description{
We obtain the filtered human breast cancers scRNA-seq dataset from 
[Zenodo data repository] (<https://doi.org/10.5281/zenodo.4739739>). 
The dataset contains the expression levels of 11920 genes and 306 spots.
}
\examples{
data(breast.st)

}
\keyword{datasets}
