% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnDeconv.R
\name{data_process}
\alias{data_process}
\title{This function focuses on cleaning scRNA-seq and stRNA-seq datasets.}
\usage{
data_process(
  sc_exp,
  sc_label,
  spot_exp,
  spot_loc,
  gene_det_in_min_cells_per = 0.01,
  expression_threshold = 1,
  nUMI = 100,
  verbose = FALSE,
  plot = FALSE
)
}
\arguments{
\item{sc_exp}{scRNA-seq matrix, genes * cells. The format should be raw-counts. The matrix need include gene names and cell names.}

\item{sc_label}{cell type information. The cells are need be divided into multiple category.}

\item{spot_exp}{stRNA-seq matrix, genes * spots. The format should be raw counts. The matrix need include gene names and spot names.}

\item{spot_loc}{coordinate matrix, spots * coordinates. The matrix need include spot names and coordinate name (x, y).}

\item{gene_det_in_min_cells_per}{a floor variable. minimum percent of genes that need to be detected in a cell.}

\item{expression_threshold}{a floor variable. Threshold to consider a gene expressed.}

\item{nUMI}{a floor variable.     minimum of read count that need to be detected in a cell or spot.}

\item{verbose}{a logical variable that defines whether to print the processing flow of data process.}

\item{plot}{a logical variable that defines whether to plot the selected genes and selected cell expression.}
}
\value{
a list includes processed scRNA-seq matrix, cell type, stRNA-seq matrix.
}
\description{
This function focuses on cleaning scRNA-seq and stRNA-seq datasets.
}
\examples{
data("breast.sc.ref")
data("breast.sc.cell.label")
data("breast.st")
data("breast.st.loc")
database <- data_process(breast.sc.ref, breast.sc.cell.label, breast.st, breast.st.loc)

}
