% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnDeconv.R
\name{EnDecon_individual_methods}
\alias{EnDecon_individual_methods}
\title{Running each base deconvolution method individually to obtain the  base cell type deconvolution results on spatially resolved transcriptomics data.}
\usage{
EnDecon_individual_methods(
  sc_exp,
  sc_label,
  spot_exp,
  spot_loc,
  gene_det_in_min_cells_per = 0.01,
  expression_threshold = 1,
  nUMI = 100,
  verbose = FALSE,
  plot = FALSE,
  python_env = NULL,
  use_gpu = FALSE,
  saving_results = FALSE,
  SCDC = TRUE,
  RCTD = TRUE,
  MuSiC = TRUE,
  DeconRNASeq = TRUE,
  DestVI = TRUE,
  DWLS = TRUE,
  SPOTlight = TRUE,
  SpatialDWLS = TRUE,
  Stereoscope = TRUE,
  cell2location = TRUE,
  CARD = TRUE,
  STdeconvolve = TRUE,
  SCDC.iter.max = 1000,
  RCTD.CELL_MIN_INSTANCE = 10,
  MuSiC.iter.max = 1000,
  MuSiC.nu = 1e-04,
  MuSiC.eps = 0.01,
  DeconRNASeq.perc = 0.05,
  DWLS.parallel = TRUE,
  DWLS.is_select_DEGs = TRUE,
  SPOTlight.cl_n = 100,
  SPOTlight.hvg = 3000,
  SPOTlight.min_cont = 0.001,
  SpatialDWLS.findmarker_method = "gini",
  SpatialDWLS.ncp_spa = 100,
  SpatialDWLS.dimensions_to_use = 10,
  SpatialDWLS.k = 10,
  SpatialDWLS.resolution = 0.4,
  SpatialDWLS.n_iterations = 1000,
  SpatialDWLS.n_cell = 50,
  SpatialDWLS.is_select_DEGs = TRUE,
  Stereoscope.sc_training_plot = FALSE,
  Stereoscope.sc_training_save_trained_model = FALSE,
  Stereoscope.sc_max_epochs = 10000,
  Stereoscope.sc_lr = 0.01,
  Stereoscope.select_HVG = TRUE,
  Stereoscope.HVG_num = 5000,
  Stereoscope.st_training_plot = FALSE,
  Stereoscope.st_training_save_trained_model = FALSE,
  Stereoscope.st_max_epochs = 10000,
  Stereoscope.st_lr = 0.01,
  cell2location.sc_max_epoches = 1000,
  cell2location.sc_lr = 0.002,
  cell2location.st_N_cells_per_location = 30,
  cell2location.st_detection_alpha = 200,
  cell2location.st_max_epoches = 10000,
  CARD.minCountGene = 100,
  CARD.minCountSpot = 5,
  STdeconvolve.min.lib.size = 100,
  STdeconvolve.min.reads = 1,
  STdeconvolve.nTopOD = 1000,
  STdeconvolve.betaScale = 1000,
  DestVI.n_top_genes = 2000,
  DestVI.max_iter_sc = 400,
  DestVI.max_iter_st = 3000
)
}
\arguments{
\item{sc_exp}{scRNA-seq matrix, genes * cells. The format should be raw-counts. The matrix need include gene names and cell names.}

\item{sc_label}{cell type information. The cell need be divided into multiple categories.}

\item{spot_exp}{stRNA-seq matrix, genes * spots. The format should be raw-counts. The matrix need include gene names and spot names.}

\item{spot_loc}{coordinate matrix, spots * coordinates. The matrix need include spot names and coordinate name (x, y).}

\item{gene_det_in_min_cells_per}{a floor variable. minimum percent # of genes that need to be detected in a cell.}

\item{expression_threshold}{a floor variable. Threshold to consider a gene expressed.}

\item{nUMI}{a floor variable.     minimum # of read count that need to be detected in a cell or spot.}

\item{verbose}{a logical variable that defines whether to print the processing flow of data process.}

\item{plot}{a logical variable that defines whether to plot the selected genes and selected cell expression.}

\item{python_env}{the path of python environment. We recommend user construct python environment by the .yml provided by ours.}

\item{use_gpu}{a logical variable whether to use GPU to train Stereoscope and cell2location.}

\item{saving_results}{a logical variable whether to save the results of individual deconvolution methods.}

\item{SCDC}{a logical variable whether to apply SCDC.}

\item{RCTD}{a logical variable whether to apply RCTD.}

\item{MuSiC}{a logical variable whether to apply MuSiC all gene and MuSiC weighted.}

\item{DeconRNASeq}{a logical variable whether to apply DeconRNASeq.}

\item{DestVI}{a logical variable whether to apply DestVI.}

\item{DWLS}{a logical variable whether to apply DWLS and SVR.}

\item{SPOTlight}{a logical variable whether to apply SPOTlight.}

\item{SpatialDWLS}{a logical variable whether to apply SpatialDWLS.}

\item{Stereoscope}{a logical variable whether to apply Stereoscope.}

\item{cell2location}{a logical variable whether to apply cell2location.}

\item{CARD}{a logical variable whether to apply CARD.}

\item{STdeconvolve}{a logical variable whether to apply STdeconvolve.}

\item{SCDC.iter.max}{a integer variable represents the maximum number of iteration in WNNLS of SCDC.}

\item{RCTD.CELL_MIN_INSTANCE}{a integer value represent the min cells in one cell type for reference scRAN-seq.}

\item{MuSiC.iter.max}{a integer variable represents maximum iteration number of MuSiC training.}

\item{MuSiC.nu}{a floor variable represents regulation parameter in MuSiC model.}

\item{MuSiC.eps}{a floor variable represents threshold of convergence of training model.}

\item{DeconRNASeq.perc}{a floor variable represents the values for filter cells.}

\item{DWLS.parallel}{a logical variable indicating whether to apply DWL with multiple CPU. Default setting is TRUE.}

\item{DWLS.is_select_DEGs}{a logical variable indicating whether to select genes for each cell type of scRNA-seq dataset. Default setting is TRUE.}

\item{SPOTlight.cl_n}{integer variable  indicating how many cells to keep from each cluster. If a cluster has n < cl_n then all cells will be selected, if it has more then cl_n will be sampled randomly. Default value is 100.}

\item{SPOTlight.hvg}{integer variable that represents number of highly variable genes to use on top of the marker genes. Default values is 3000.}

\item{SPOTlight.min_cont}{floor variable indicates the minimum contribution we expect from a cell in that spot. Default values is 0.001.}

\item{SpatialDWLS.findmarker_method}{a string vector indicating method to use to detect differentially expressed genes.}

\item{SpatialDWLS.ncp_spa}{a integer value indicating number of principal components to calculate. Default setting is 100.}

\item{SpatialDWLS.dimensions_to_use}{a integer value indicating number of dimensions to use as input for constructing KNN network. Default setting is 10.}

\item{SpatialDWLS.k}{a integer value indicating number of k neighbors to use for constructing KNN network. Default setting is 10.}

\item{SpatialDWLS.resolution}{resolution in doLeidenCluster function in Giotto package. Default setting is 0.4.}

\item{SpatialDWLS.n_iterations}{number of interations to run the Leiden algorithm. If the number of iterations is negative, the Leiden algorithm is run until an iteration in which there was no improvement.}

\item{SpatialDWLS.n_cell}{number of cells per spot. Default setting is 50.}

\item{SpatialDWLS.is_select_DEGs}{a logical value whether to select genes before applying for the SpatialDWLS.}

\item{Stereoscope.sc_training_plot}{a logical variable whether to plot the training loss indicating whether to increase the number of maximum epoch for training for scRNA-seq dataset. Default setting is FALSE.}

\item{Stereoscope.sc_training_save_trained_model}{a logical variable whether to save the trained model for scRNA-seq dataset. Default setting is FALSE.}

\item{Stereoscope.sc_max_epochs}{an integer variable indicating the maximum epoches for training scRNA-seq. Default setting is 400.}

\item{Stereoscope.sc_lr}{an integer variable indicating the learning rate for training scRNA-seq. Default setting is 0.01.}

\item{Stereoscope.select_HVG}{a logical variable whether to select highly variable genes for the scRNA-seq data. Default setting is TRUE.}

\item{Stereoscope.HVG_num}{number of selected highly variable genes if Stereoscope.select_HVG = TRUE. Default setting is 5000.}

\item{Stereoscope.st_training_plot}{a logical variable whether to plot the training loss indicating whether to increase the number of maximum epoch for training for stRNA-seq dataset. Default setting is FALSE.}

\item{Stereoscope.st_training_save_trained_model}{a logical variable whether to plot the training loss indicating whether to increase the number of maximum epoch for training for stRNA-seq dataset. Default setting is FALSE.}

\item{Stereoscope.st_max_epochs}{an integer variable indicating the maximum epoches for training sTRNA-seq. Default setting is 400.}

\item{Stereoscope.st_lr}{an integer variable indicating the learning rate for training sTRNA-seq. Default setting is 0.01.}

\item{cell2location.sc_max_epoches}{an integer variable indicating the maximum epoches for training scRNA-seq.}

\item{cell2location.sc_lr}{an integer variable indicating the learning rate for training scRNA-seq.}

\item{cell2location.st_N_cells_per_location}{a integer variable indicating the number of cells in each spot.}

\item{cell2location.st_detection_alpha}{a floor variable indicating the super-parameter of regularization.}

\item{cell2location.st_max_epoches}{an integer variable indicating the maximum epoches for training stRNA-seq.}

\item{CARD.minCountGene}{an integer variable indicating the minimum counts for each gene for the construct CARD object. Default setting is 100.}

\item{CARD.minCountSpot}{an integer variable indicating the minimum counts for each spatial location. Default setting is 5.}

\item{STdeconvolve.min.lib.size}{Minimum number of genes detected in a cell. Cells with fewer genes will be removed for the stRNA-seq dataset. Default setting is 1.}

\item{STdeconvolve.min.reads}{Minimum number of reads per gene. Genes with fewer reads will be removed. Default setting is 1.}

\item{STdeconvolve.nTopOD}{Number of top over-dispersed genes to use for the stRNA-seq data. Default setting is 1000.}

\item{STdeconvolve.betaScale}{Factor to scale the predicted cell-type gene expression profiles. Default setting is 1000.}

\item{DestVI.n_top_genes}{Number of selected HVGs by Seurat.V3. Default setting is 2000.}

\item{DestVI.max_iter_sc}{Maximum number of epoches for the training of scRNA-seq data. Default setting is 400.}

\item{DestVI.max_iter_st}{Maximum number of epoches for the training of stRNA-seq data. Default setting is 3000.}
}
\value{
a list contains all the results inferred by individual deconvolution methods and the times of running individual methods. The elements of list is a matrix, spots * cell-type and a time vector.
}
\description{
This function is implemented to perform individual deconvolution methods.  The current implementation of
EnDecon integrates twelve state-of-the-art methods:  CARD, cell2location, DeconRNASeq, DWLS, MuSiC (MuSiC weighted and MuSiC all gene), RCTD, SCDC, SpatialDWLS,
SPOTlight,Stereoscope, and SVR. These packages will be automatically installed along
with EnDecon. \cr
}
