% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MVC.reference.cell.label}
\alias{MVC.reference.cell.label}
\title{cell type labels of reference cells for the third scenario in our simulation study.}
\format{
a vector
}
\usage{
data(MVC.reference.cell.label)
}
\description{
cell type labels of reference cells for the third scenario in our simulation study.
}
\examples{
data(MVC.reference)

}
\keyword{datasets}
