% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MVC.reference}
\alias{MVC.reference}
\title{raw count matrix of reference scRNA-seq dataset}
\format{
a large matrix
}
\usage{
data(MVC.reference)
}
\description{
We obtain the mouse primary visual cortex (VISp) scRNA-seq dataset from Smart-seq protocol.
The raw count dataset contains the expression levels of 981 genes in 10549 cells. For details, please refer our article.
}
\examples{
data(MVC.reference)

}
\keyword{datasets}
