% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MVC.ST.coor}
\alias{MVC.ST.coor}
\title{coordinate of spots for the spatially resolved transcriptomics data for the third scenario in our simulation study.}
\format{
a list
}
\usage{
data(MVC.ST.coor)
}
\description{
coordinate of spots for the MVC.ST data. The center of the grids are served as the coordinates of the corresponding generated spots.
}
\examples{
data(MVC.ST.coor)

}
\keyword{datasets}
