% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MVC.ST}
\alias{MVC.ST}
\title{raw count of spatially resolved transcriptomics data for the third scenario in our simulation study.
select a public STARmap dataset, which contains the expression levels of 10,000 genes in 973 cells from the mouse visual cortex at
the single-cell resolution and is refined with six neocortical layers. To generate coarse-grained SRT data from single-cell resolution data, we define one spot-based region
 by the size of the grid and aggregate the gene expression level that fall into each spot.
 After gridding, a total of 175 spots are simulated and each spot covers 1~13 cells.}
\format{
a large matrix
}
\usage{
data(MVC.ST)
}
\description{
raw count of spatially resolved transcriptomics data for the third scenario in our simulation study.
select a public STARmap dataset, which contains the expression levels of 10,000 genes in 973 cells from the mouse visual cortex at
the single-cell resolution and is refined with six neocortical layers. To generate coarse-grained SRT data from single-cell resolution data, we define one spot-based region
 by the size of the grid and aggregate the gene expression level that fall into each spot.
 After gridding, a total of 175 spots are simulated and each spot covers 1~13 cells.
}
\examples{
data(MVC.ST)

}
\keyword{datasets}
