% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCTD_pipeline.R
\name{RCTD_pipeline}
\alias{RCTD_pipeline}
\title{This function is to deconvolute ST data using the RCTD method}
\usage{
RCTD_pipeline(sc_data, st_data)
}
\arguments{
\item{sc_data}{this is a seurat object, assay = "RNA"}

\item{st_data}{this is a seurat object, assay = "Spatial"}
}
\value{
This function returns the result of RCTD deconvolution
}
\description{
This function is to deconvolute ST data using the RCTD method
}
