% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnDeconv.R
\name{solve_ensemble}
\alias{solve_ensemble}
\title{Ensemble the results of individual deconvolution results.
This function uses the weighted median methods to integrate the results obtained by individual deconvolution methods.}
\usage{
solve_ensemble(
  Results.Deconv,
  lambda = NULL,
  prob.quantile = 0.5,
  niter = 100,
  epsilon = 1e-05
)
}
\arguments{
\item{Results.Deconv}{a list contains all the results of individual deconvolution methods. The elements of list is a matrix, spots * cell-type.}

\item{lambda}{hyper-parameter constrain the weight of individual methods for ensemble. If the parameter is set to NULL, then, we will adopt the value in our algorithm.}

\item{prob.quantile}{numeric of probabilities with values in [0,1]. Default setting is 0.5.}

\item{niter}{a positive integer represents the maximum number of updating algorithm. Default setting is 100.}

\item{epsilon}{a parameter represents the stop criterion.}
}
\value{
a list contains a matrix of the ensemble deconvolution result and a vector of the weight assigned to individual methods.
}
\description{
Ensemble the results of individual deconvolution results.
This function uses the weighted median methods to integrate the results obtained by individual deconvolution methods.
}
\examples{
data("MVC.reference")
data("MVC.reference.cell.label")
data("MVC.ST")
data("MVC.ST.coor")
##### path on ubuntu platform on our computer
python_env <- "~/.conda/envs/EnDecon_GPU/bin/python"
##### we use all the genes for the deconvolutioin
Results.Deconv <- EnDecon_individual_methods(MVC.reference, MVC.reference.cell.label,
MVC.ST, MVC.ST.coor,  python_env = python_env, use_gpu = TRUE,
RCTD.CELL_MIN_INSTANCE = 10, gene_det_in_min_cells_per = 0,
expression_threshold = 0, nUMI = 1, DWLS.is_select_DEGs = FALSE,
SpatialDWLS.is_select_DEGs = FALSE)
ensemble.results <- solve_ensemble(Results.Deconv)

}
